refactor: 함수 분리 및 DateSpinnerListener 클래스 분리
